improvements/enhanced_requirements.txt
# Core ML and Data Processing
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# OCR and Image Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=10.0.0

# Document Processing
python-docx>=0.8.11
pdf2image>=1.16.3
PyPDF2>=3.0.0

# Web Interface
streamlit>=1.28.0
plotly>=5.17.0

# API and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and Storage
redis>=5.0.0

# Advanced ML Models (Optional)
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
aiofiles>=23.2.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Cloud Integration (Optional)
boto3>=1.34.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0
